Make filter icon animation smoother
With these changes, the animation won't stop abruptly.

Previously clicking on the filter multiple times or clicking on another
filter while the animation is playing can cause the animation to end
abruptly. Now the animation always plays to the end before it can start
again.

The animation length is 700ms, so the timeout was made shorter.